+++ PapaCheck.Web/js/big-screen.js
@@ -1183,7 +1183,6 @@ async function redeemItem(itemId) {
     showToast('兑换成功！');
     Voice.speak('兑换成功！');
     Voice.speak('奖励箱有新奖励，快去看看吧');
-    window._rewardBoxVoiceHandled = true;
   } finally {
     _redeemingItem = false;
   }